Removing sequencer cross-class dependancy
Removes the extern function show_current_step from sequencer. The function always is called for the playingPattern which won't change in the sequencer functions so no need to have sequencer update the drawing for this.

The sequencer_mode will now call this function after calling doStep functions.

Renamed omxLeds.updateLeds to updateBlinkStates which describes the function better.