Fixes arp getting having stuck notes if chance is lower than 100
Problem was that if chance kicked in and the notes did not get sent to the arp on a noteon event, then when the noteoff event happend, if it got sent to arp instead of passing through, a note off would not be sent for the note and a single note will get stuck on. The arp could also get stuck, acting like it is in hold mode.

To fix this, I tracked notes getting passed through using trackingNoteGroupsPassthrough for up to 8 notes. notes with fixed lengths ignored since they will turn off automatically.

By tracking these notes, I pass them through for noteoffs in the case they were previously passed through noteons.